requirements.txt
# requirements.txt
# ============================================================================
# Flask & Web
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# LangChain & RAG
langchain==0.2.12
langchain-community==0.2.10
langchain-text-splitters==0.2.2

# Vector Store & Embeddings
faiss-cpu
sentence-transformers==2.2.2

# Document Loaders
pypdf==3.17.0
python-docx==1.1.0
unstructured==0.11.2
openpyxl==3.1.2

# LLM
groq==0.4.2

# Audio
openai-whisper==20231117

# Utilities
tenacity==8.2.3
python-dotenv==1.0.0